requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
docopt==0.6.2
idna==3.3
MangaDex.py==2.0.10
Pillow==10.0.0
pipreqs==0.4.11
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
yarg==0.1.9
